% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_and_add_2.R
\name{addition}
\alias{addition}
\title{Add some columns back to a particle}
\usage{
addition(particle, GB, left)
}
\arguments{
\item{particle}{a matrix representing a particle.}

\item{GB}{a matrix representing a global best particle.}

\item{left}{the \code{left} return from \code{sibma::deletion}.}
}
\description{
Add some columns back to a particle
}
