% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{all_equal_list}
\alias{all_equal_list}
\title{Check All Elements in a List are Identical}
\usage{
all_equal_list(a_list)
}
\arguments{
\item{a_list}{a list.}
}
\description{
Check All Elements in a List are Identical
}
