% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cl.R
\name{c_particle_orth}
\alias{c_particle_orth}
\title{Transformation: a factor's design to corresponding effects}
\usage{
c_particle_orth(col_particle, factor_level)
}
\arguments{
\item{col_particle}{an integer vector specifying a particle's column.}

\item{factor_level}{an integer vector specifying col_particle's levels.}
}
\description{
Transform a particle's column, which represents a factor's design to its
corresponding effects, such as a linear effect for a two-level factor or
a linear effect and a quadratic effect for a three-level factor.
}
