% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence
Check if the word-length patterns calculated from each SIB process are the same.}
\usage{
check_convergence(MA_df)
}
\arguments{
\item{MA_df}{a list. Each element is a list describing the word-length patterns in
each SIB process.}
}
\description{
Check Convergence
Check if the word-length patterns calculated from each SIB process are the same.
}
