% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlength_pattern_2.R
\name{compare_MA_min}
\alias{compare_MA_min}
\title{Compare multiple word-length pattern}
\usage{
compare_MA_min(particle, MA)
}
\arguments{
\item{particle}{a list containing all particles.}

\item{MA}{a list containing all the word-length patterns of their
corresponding particles.}
}
\description{
Compare a list of word-length pattern and yield the minimum aberration design.
}
\details{
The function will return three values: \code{best}, \code{result}, \code{GB}.
\code{best} returns a vector of integers indicating where  best particles are.
\code{result} returns a matrix. Each row is a word-length pattern.
\code{GB} returns a list of matrices specifying the best particles.
}
