% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cl.R
\name{create_particle}
\alias{create_particle}
\title{Create a list of particles}
\usage{
create_particle(
  factor_level,
  unit,
  particle_number,
  incidence_matrix,
  treatment_effect
)
}
\arguments{
\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{unit}{an integer describing the number of experimental units to be used
in a particle.}

\item{particle_number}{an integer describing the number of particles.}

\item{incidence_matrix}{a list. Each element is a matrix specifying an
incidence matrix.}

\item{treatment_effect}{a list. Each element is a numeric vector specifying
which factors should have the same treatment effect in terms of a certain
incidence matrix. The order of elements in \code{treatment_effect} should be
corresponding to that in \code{incidence_matrix}.}
}
\description{
Each particle is a kind of experimental design where the number of rows
represent the number of experimental units and the number of columns represent
the number of factors.
}
\details{
We highly recommend that you use \code{c(-1,1)} as a code for a 2-level
factor. If you have a factor with p-level, we recommend you use \code{0:(p-1)}
as the code for this factor.
}
