% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_and_add.R
\name{deletion}
\alias{deletion}
\title{Delete some columns within a particle}
\usage{
deletion(particle, all_two_level, factor_level, P_w, q, total_unit, unit)
}
\arguments{
\item{particle}{a matrix representing a particle.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{q}{an integer describing how many columns should be deleted.}

\item{total_unit}{an integer representing the number of total run size in a
full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used
in a particle.}
}
\description{
Delete some columns within a particle
}
