% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{evaluation}
\alias{evaluation}
\title{Update Local best or Global Best Particles}
\usage{
evaluation(
  X,
  LB_0,
  LB,
  evaluate_GB,
  factor_number,
  P_w,
  all_two_level,
  factor_level,
  total_unit,
  unit,
  multiply_len
)
}
\arguments{
\item{X}{a list. Each element is a particle after \code{sibma::move()}.}

\item{LB_0}{representing the current local best particles of \code{X}.}

\item{LB}{a list representing all \code{LB} particles.}

\item{evaluate_GB}{logical. If you want to update the \code{GB} particle,
\code{evaluate_GB} should be \code{TRUE}; otherwise, it should be
\code{FALSE}.}

\item{factor_number}{an integer describing the number of factors.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{total_unit}{an integer representing the number of total run size in
a full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used in a
particle.}

\item{multiply_len}{a numeric vector. Each element is used to modify
the value of each column in a model matrix.}
}
\description{
Update Local best or Global Best Particles
}
