% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{initial}
\alias{initial}
\title{Initial step}
\usage{
initial(
  particle,
  all_two_level,
  P_w,
  factor_level,
  total_unit,
  unit,
  multiply_len
)
}
\arguments{
\item{particle}{a list containing all particles.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{total_unit}{an integer representing the number of total run size in a
full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used in a
particle.}

\item{multiply_len}{a numeric vector. Each element is used to modify
the value of each column in a model matrix.}
}
\description{
Initial step
}
