% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{MA_ordering}
\alias{MA_ordering}
\title{Return the Order of a Data Frame of Word-Length Patterns}
\usage{
MA_ordering(MA_df)
}
\arguments{
\item{MA_df}{a data frame containing word-length patterns.}
}
\description{
Return the Order of a Data Frame of Word-Length Patterns
}
