% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_2.R
\name{mix_with_GB}
\alias{mix_with_GB}
\title{Mix with Global Best Particles}
\usage{
mix_with_GB(X, all_two_level, GB, P_w, q_GB, factor_level, total_unit, unit)
}
\arguments{
\item{X}{a list. Each element is a particle ready to be mixed.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{GB}{a list. Each element is a global best particle.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{q_GB}{an integer describing how many columns of each element of\code{X}
should be mixed with the corresponding columns of each element of \code{GB}.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{total_unit}{an integer representing the number of total run size in a
full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used
in a particle.}
}
\description{
Mix with Global Best Particles
}
