% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix_with_new}
\alias{mix_with_new}
\title{Mix with a New Particle}
\usage{
mix_with_new(
  X,
  all_two_level,
  P_w,
  q_new,
  factor_level,
  total_unit,
  unit,
  incidence_matrix,
  treatment_effect
)
}
\arguments{
\item{X}{a list. Each element is a particle ready to be mixed.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{q_new}{an integer describing how many columns of each element of\code{X}
should be mixed with the corresponding columns of a new particle created from
\code{sibma::create_particle()}.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{total_unit}{an integer representing the number of total run size in a
full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used
in a particle.}

\item{incidence_matrix}{a list. Each element is a matrix specifying an
incidence matrix.}

\item{treatment_effect}{a list. Each element is a numeric vector specifying
which factors should have the same treatment effect in terms of a certain
incidence matrix. The order of elements in \code{treatment_effect} should be
corresponding to that in \code{incidence_matrix}.}
}
\description{
Mix with a New Particle
}
