% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Update Current Particles}
\usage{
move(
  X,
  mixwGB,
  mixwLB,
  t_1,
  P_w,
  q_new,
  all_two_level,
  factor_level,
  total_unit,
  unit,
  multiply_len,
  incidence_matrix,
  treatment_effect
)
}
\arguments{
\item{X}{a list. Each element is a particle ready to be mixed.}

\item{mixwGB}{a list. Each element is a particle after \code{X} mixed with
\code{GB}.}

\item{mixwLB}{a list. Each element is a particle after \code{X} mixed with
\code{LB}.}

\item{t_1}{logical. If this is the first move step, \code{t_1} should be
\code{TRUE}; otherwise, it should be \code{FALSE}.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{q_new}{an integer describing how many columns of each element of\code{X}
should be mixed with the corresponding columns of a new particle created from
\code{sibma::create_particle()}.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{total_unit}{an integer representing the number of total run size in
a full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used in a
particle.}

\item{multiply_len}{a numeric vector. Each element is used to modify
the value of each column in a model matrix.}

\item{incidence_matrix}{a list. Each element is a matrix specifying an
incidence matrix.}

\item{treatment_effect}{a list. Each element is a numeric vector specifying
which factors should have the same treatment effect in terms of a certain
incidence matrix. The order of elements in \code{treatment_effect} should be
corresponding to that in \code{incidence_matrix}.}
}
\description{
Update Current Particles
}
