% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cl.R
\name{particle_orth}
\alias{particle_orth}
\title{Transformation: All particles to corresponding effects}
\usage{
particle_orth(particle, factor_level)
}
\arguments{
\item{particle}{a list containing all particles.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}
}
\description{
Transform every particle, namely, every experimental design to its corresponding
effects.
}
