% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxty_2.R
\name{sxty_SIB_relay_only_t}
\alias{sxty_SIB_relay_only_t}
\title{Conduct multiple SIB with Increased Number of Iteration}
\usage{
sxty_SIB_relay_only_t(
  prior_SIB,
  history,
  factor_level,
  SIB_time,
  all_two_level,
  P_w,
  q_GB,
  q_LB,
  q_new,
  t,
  total_unit,
  multiply_len,
  incidence_matrix,
  treatment_effect
)
}
\arguments{
\item{prior_SIB}{the \code{testSIB} return of the previous
\code{sxty_SIB} method.}

\item{history}{the \code{history} return of the previous
\code{sxty_SIB} method.}

\item{factor_level}{a list. Each element is a numeric vector specifying
levels of a factor.}

\item{SIB_time}{an integer indicating the number of SIB process you would
like to run.}

\item{all_two_level}{logical. If all factors are two levels,
\code{all_two_level} should be \code{TRUE}; otherwise it should be
\code{FALSE}.}

\item{P_w}{a list. Each element is a matrix describing the orthogonal
projection matrix onto the corresponding stratum variance.}

\item{q_GB}{an integer describing how many columns of a particle
should be mixed with the corresponding columns of \code{GB}.}

\item{q_LB}{an integer describing how many columns of a particle
should be mixed with the corresponding columns of \code{LB}.}

\item{q_new}{an integer describing how many columns of a particle
should be mixed with the corresponding columns of a random
particle created by \code{sibma::create_particle()}.}

\item{t}{an integer indicating the increased number of iteration.}

\item{total_unit}{an integer representing the number of total run size in
a full factorial design.}

\item{multiply_len}{a numeric vector. Each element is used to modify
the value of each column in a model matrix.}

\item{incidence_matrix}{a list. Each element is a matrix specifying an
incidence matrix.}

\item{treatment_effect}{a list. Each element is a numeric vector specifying
which factors should have the same treatment effect in terms of a certain
incidence matrix. The order of elements in \code{treatment_effect} should be
corresponding to that in \code{incidence_matrix}.}
}
\description{
Conduct multiple SIB with Increased Number of Iteration
}
