% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlength_pattern_2.R
\name{worlen_pattern}
\alias{worlen_pattern}
\title{Create wordlength pattern}
\usage{
worlen_pattern(factor_number, U, P_w, total_unit, unit)
}
\arguments{
\item{factor_number}{an integer describing the number of factors.}

\item{U}{a matrix describing a model matrix calculated from
\code{sibma::general_create_U()}.}

\item{P_w}{a matrix describing the orthogonal projection matrix onto the
corresponding stratum variance.}

\item{total_unit}{an integer representing the number of total run size in a
full factorial design.}

\item{unit}{an integer describing the number of experimental units to be used in a
particle.}
}
\description{
Create an experimental design's wordlength pattern.
}
